requirements.txt
PyGObject
sounddevice
numpy
webrtcvad
requests
python-dotenv
soundfile
vosk
piper-tts
kittentts
